requirements.txt
python-telegram-bot==20.3
python-bidi==0.4.2